Fixed some code, tested assignActivity

+++ src/main/java/polimi/aui/sentimentaigroup6b/entities/Audio.java
@@ -21,7 +21,7 @@ public class Audio {
     private Session session;
 
     @ElementCollection
-    private List<Float> detectedEmotions;
+    private List<Double> detectedEmotions;
 
     @Lob
     private byte[] audioData;
